// src/types/bigthree.ts
export interface BigThreeStatus {
  mostLevel: number;
  sumAverage: number;
  benchAverage: number;
  deadliftAverage: number;
  squatAverage: number;
  recentSum: number;
  recentBench: number;
  recentDeadlift: number;
  recentSquat: number;
  benchRate: number;
  deadliftRate: number;
  squatRate: number;
}